requirements.txt
Django==2.2.3
djangorestframework==3.9.4
django-filter==2.1.0
markdown
